split cmd_moderation into moderation and bookkeeping
